+++ CHANGELOG.md
@@ -10,11 +10,21 @@ and this project adheres to
 
 Back to [Readme](README.md).
 
+[1.4.0]: https://github.com/bischoffdev/property-aggregator/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/bischoffdev/property-aggregator/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/bischoffdev/property-aggregator/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/bischoffdev/property-aggregator/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/bischoffdev/property-aggregator/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/bischoffdev/property-aggregator/tree/v1.0.0
+
+## [1.4.0] - 2022-03-21
+
+### Added
+* New `withPredicate` method to specify a custom filter that is processed on 
+  the properties during the final list processing
+
+### Changed
+* More resilient handling of missing properties files in the classpath.
 
 ## [1.3.0] - 2022-03-15